Updated the onAnimeClicked callbacks to pass animeId Int

+++ feature/home/src/main/java/com/mumbicodes/home/navigation/HomeScreenNavigation.kt
@@ -26,7 +26,7 @@ fun NavController.navigateToHomeGraph(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.homeGraph(
     navController: NavHostController,
-    onAnimeClicked: () -> Unit,
+    onAnimeClicked: (Int) -> Unit,
     onSeeAllButtonClicked: () -> Unit
 ) {
     navigation(
